added basic functionality for the spells

+++ app/events/TileClicked.java
@@ -11,6 +11,9 @@ import structures.basic.Tile;
 import structures.basic.Unit;
 import structures.basic.cards.Card;
 import structures.basic.cards.Wraithling;
+import structures.basic.player.Hand;
+import structures.basic.player.HumanPlayer;
+import structures.basic.player.Player;
 
 /**
  * Indicates that the user has clicked an object on the game canvas, in this case a tile.
@@ -72,37 +75,48 @@ public class TileClicked implements EventProcessor {
 	// Process spell casting based on target tile
 	private void handleSpellCasting(GameState gameState, Card card, Tile tile) {
 	    // Spell casting logic
-	    if (card.getCardname().equals("Horn of the Forsaken")) {
-	        if (gameState.getHuman().getMana() >= card.getManacost()) {
-	            // Sufficient mana for casting the spell
-	            gameState.gameService.HornOfTheForesaken(card);//to change to fit other spell	            	            
-	            gameState.getCurrentPlayer().setRobustness(gameState.getCurrentPlayer().getRobustness() + 3);	            
-	            System.out.println("Player's robustness: " + gameState.getHuman().getRobustness());
-	            gameState.gameService.removeCardFromHandAndSummon(card, tile);
-	            
-	        } else {
-	            // Insufficient mana for casting the spell
-//	            BasicCommands.addPlayer1Notification(out, "Not enough mana", 2);
-	        	System.out.println("Not enough mana");
-	    	    gameState.gameService.removeHighlightFromAll();
-	    	    gameState.gameService.notClickingCard();
+		if (gameState.getCurrentPlayer() instanceof HumanPlayer) {
+			Player player = gameState.getCurrentPlayer();
+			Hand hand = player.getHand();
+			int handPosition = gameState.getCurrentCardPosition();
 
-	        }
-	    }
-		if (card.getCardname().equals("Dark Terminus") &&
-				tile.getUnit().getOwner() != gameState.getHuman()) {
-			    gameState.gameService.performUnitDeath(tile.getUnit());
-				try {
-					Thread.sleep(100);
-				} catch (InterruptedException e) {
-					e.printStackTrace();
-				}
-			    Wraithling.summonWraithlingToTile(tile, out, gameState);
-	    	
-	    }
-	    
-	    // Remove highlight from all tiles
-	    gameState.gameService.removeHighlightFromAll();
+		    if (card.getCardname().equals("Horn of the Forsaken")) {
+		        if (gameState.getHuman().getMana() >= card.getManacost()) {
+		            // Sufficient mana for casting the spell
+		            gameState.gameService.HornOfTheForesaken(card);//to change to fit other spell	
+		            gameState.getCurrentPlayer().setRobustness(player.getRobustness() + 3);	            
+		            System.out.println("Player's robustness: " + player.getRobustness());
+		            hand.removeCardAtPosition(handPosition);
+		    		gameState.gameService.updateHandPositions(player.getHand());
+		    		
+		        } else {
+		            // Insufficient mana for casting the spell
+		        	BasicCommands.addPlayer1Notification(out, "Not enough mana", 2);
+		        	System.out.println("Not enough mana");
+		    	    gameState.gameService.removeHighlightFromAll();
+		    	    gameState.gameService.notClickingCard();
+
+		        }
+		    }
+			if (card.getCardname().equals("Dark Terminus") &&
+					tile.getUnit().getOwner() != gameState.getHuman()) {
+				    gameState.gameService.performUnitDeath(tile.getUnit());
+				    
+					try {
+						Thread.sleep(100);
+					} catch (InterruptedException e) {
+						e.printStackTrace();
+					}
+					System.out.println("Dark Terminus is casted");
+				    Wraithling.summonWraithlingToTile(tile, out, gameState);
+		    	
+		    }
+		    
+		    // Remove highlight from all tiles
+		    gameState.gameService.removeHighlightFromAll();
+			
+		}
+
 	}
 	
 	// Process unit move or attack based on targetTile's state
